Server/requirements.txt
fastapi==0.115.5
python-dotenv==0.21.0
uvicorn==0.31.1
langchain-openai==0.2.11
pymupdf4llm==0.0.17
langchain==0.3.9
langgraph==0.2.53
langchain-community==0.3.9
sentence-transformers==3.3.1
langchain-anthropic==0.3.0
langfuse==2.57.0
pytest==7.4.4
deepeval==2.0.6